Bot now saves its state.
This is part 1 of making it safe to bring down often. It's now robust
enough that brutally killing it is pretty safe, and it will at worst
only send out some extra warnings for the last minute of posts.

Next I'm gonna make it handle a hangup signal gracefully such that it
comes down after `run_once` finishes up.